+++ blaze-server/src/test/scala/org/http4s/server/blaze/BlazeServerSuite.scala
@@ -65,7 +65,7 @@ class BlazeServerSuite extends Http4sSuite {
       .withHttpApp(service)
       .resource
 
-  def blazeServer: FunFixture[Server[IO]] =
+  val blazeServer =
     ResourceFixture[Server[IO]](
       serverR,
       (_: TestOptions, _: Server[IO]) => IO.unit,
@@ -120,16 +120,16 @@ class BlazeServerSuite extends Http4sSuite {
     }
 
   blazeServer.test("route requests on the service executor") { server =>
-    get(server, "/thread/routing").map(_.startsWith("http4s-spec-")).assertEquals(true)
+    get(server, "/thread/routing").map(_.startsWith("http4s-spec-")).assert
   }
 
   blazeServer.test("execute the service task on the service executor") { server =>
-    get(server, "/thread/effect").map(_.startsWith("http4s-spec-")).assertEquals(true)
+    get(server, "/thread/effect").map(_.startsWith("http4s-spec-")).assert
   }
 
   blazeServer.test("be able to echo its input") { server =>
     val input = """{ "Hello": "world" }"""
-    post(server, "/echo", input).map(_.startsWith(input)).assertEquals(true)
+    post(server, "/echo", input).map(_.startsWith(input)).assert
   }
 
   blazeServer.test("return a 503 if the server doesn't respond") { server =>