+++ ember-server/src/main/scala/org/http4s/ember/server/internal/ServerHelpers.scala
@@ -16,8 +16,8 @@
 
 package org.http4s.ember.server.internal
 
+import cats._
 import fs2._
-import fs2.concurrent._
 import fs2.io.tcp._
 import fs2.io.tls._
 import cats.effect._
@@ -29,10 +29,9 @@ import org.http4s._
 import org.http4s.implicits._
 import org.http4s.headers.{Connection, Date}
 import _root_.org.http4s.ember.core.{Encoder, Parser}
-import _root_.org.http4s.ember.core.Util.readWithTimeout
 import _root_.io.chrisdavenport.log4cats.Logger
 import cats.data.NonEmptyList
-
+import _root_.org.http4s.ember.core.Util.durationToFinite
 private[server] object ServerHelpers {
 
   private val closeCi = "close".ci
@@ -40,6 +39,9 @@ private[server] object ServerHelpers {
   private val connectionCi = "connection".ci
   private val close = Connection(NonEmptyList.of(closeCi))
   private val keepAlive = Connection(NonEmptyList.one("keep-alive".ci))
+
+  private val serverFailure =
+    Response(Status.InternalServerError).putHeaders(org.http4s.headers.`Content-Length`.zero)
 
   def server[F[_]: ContextShift](
       bindAddress: InetSocketAddress,
@@ -49,109 +51,16 @@ private[server] object ServerHelpers {
       ready: Deferred[F, Either[Throwable, Unit]],
       shutdown: Shutdown[F],
       // Defaults
-      onError: Throwable => Response[F] = { (_: Throwable) =>
-        Response[F](Status.InternalServerError)
-      },
+      errorHandler: Throwable => F[Response[F]],
       onWriteFailure: (Option[Request[F]], Response[F], Throwable) => F[Unit],
-      maxConcurrency: Int = Int.MaxValue,
-      receiveBufferSize: Int = 256 * 1024,
-      maxHeaderSize: Int = 10 * 1024,
-      requestHeaderReceiveTimeout: Duration = 5.seconds,
-      idleTimeout: Duration = 60.seconds,
+      maxConcurrency: Int,
+      receiveBufferSize: Int,
+      maxHeaderSize: Int,
+      requestHeaderReceiveTimeout: Duration,
+      idleTimeout: Duration,
       additionalSocketOptions: List[SocketOptionMapping[_]] = List.empty,
       logger: Logger[F]
-  )(implicit F: Concurrent[F], C: Clock[F]): Stream[F, Nothing] = {
-    def socketReadRequest(
-        socket: Socket[F],
-        requestHeaderReceiveTimeout: Duration,
-        receiveBufferSize: Int,
-        isReused: Boolean
-    ): F[Request[F]] = {
-      val (initial, readDuration) = (requestHeaderReceiveTimeout, idleTimeout, isReused) match {
-        case (fin: FiniteDuration, idle: FiniteDuration, true) => (true, idle + fin)
-        case (fin: FiniteDuration, _, false) => (true, fin)
-        case _ => (false, Duration.Zero)
-      }
-
-      SignallingRef[F, Boolean](initial).flatMap { timeoutSignal =>
-        C.realTime(MILLISECONDS)
-          .flatMap(now =>
-            Parser.Request
-              .parser(maxHeaderSize)(
-                readWithTimeout[F](socket, now, readDuration, timeoutSignal.get, receiveBufferSize)
-              )
-              .flatMap { req =>
-                timeoutSignal.set(false).as(req)
-              })
-      }
-    }
-
-    def upgradeSocket(
-        socketInit: Socket[F],
-        tlsInfoOpt: Option[(TLSContext, TLSParameters)]): Resource[F, Socket[F]] =
-      tlsInfoOpt.fold(socketInit.pure[Resource[F, *]]) { case (context, params) =>
-        context
-          .server(socketInit, params, { (s: String) => logger.trace(s) }.some)
-          .widen[Socket[F]]
-      }
-
-    def runApp(socket: Socket[F], isReused: Boolean): F[(Request[F], Response[F])] =
-      for {
-        req <- socketReadRequest(socket, requestHeaderReceiveTimeout, receiveBufferSize, isReused)
-        resp <- httpApp.run(req).handleError(onError)
-      } yield (req, resp)
-
-    def send(socket: Socket[F])(request: Option[Request[F]], resp: Response[F]): F[Unit] =
-      Encoder
-        .respToBytes[F](resp)
-        .through(socket.writes())
-        .compile
-        .drain
-        .attempt
-        .flatMap {
-          case Left(err) => onWriteFailure(request, resp, err)
-          case Right(()) => Sync[F].pure(())
-        }
-
-    def postProcessResponse(req: Request[F], resp: Response[F]): F[Response[F]] = {
-      val reqHasClose = req.headers.exists {
-        // We know this is raw because we have not parsed any headers in the underlying alg.
-        // If Headers are being parsed into processed for in ParseHeaders this is incorrect.
-        case Header.Raw(name, values) => name == connectionCi && values.contains(closeCi.value)
-        case _ => false
-      }
-      val connection: Connection =
-        if (reqHasClose) close
-        else keepAlive
-      for {
-        date <- HttpDate.current[F].map(Date(_))
-      } yield resp.withHeaders(Headers.of(date, connection) ++ resp.headers)
-    }
-
-    def withUpgradedSocket(socket: Socket[F]): Stream[F, Nothing] =
-      (Stream(false) ++ Stream(true).repeat)
-        .flatMap { isReused =>
-          Stream
-            .eval(runApp(socket, isReused).attempt)
-            .evalMap {
-              case Right((req, resp)) =>
-                postProcessResponse(req, resp).map(resp => (req, resp).asRight[Throwable])
-              case other => other.pure[F]
-            }
-            .evalTap {
-              case Right((request, response)) => send(socket)(Some(request), response)
-              case Left(err) => send(socket)(None, onError(err))
-            }
-        }
-        .takeWhile {
-          case Left(_) => false
-          case Right((req, resp)) =>
-            !(
-              req.headers.get(Connection).exists(_.hasClose) ||
-                resp.headers.get(Connection).exists(_.hasClose)
-            )
-        }
-        .drain
+  )(implicit F: Concurrent[F], T: Timer[F]): Stream[F, Nothing] = {
 
     val server: Stream[F, Resource[F, Socket[F]]] =
       Stream
@@ -167,10 +76,139 @@ private[server] object ServerHelpers {
       .map { connect =>
         shutdown.trackConnection >>
           Stream
-            .resource(connect.flatMap(upgradeSocket(_, tlsInfoOpt)))
-            .flatMap(withUpgradedSocket(_))
+            .resource(connect.flatMap(upgradeSocket(_, tlsInfoOpt, logger)))
+            .flatMap(
+              runConnection(
+                _,
+                logger,
+                idleTimeout,
+                receiveBufferSize,
+                maxHeaderSize,
+                requestHeaderReceiveTimeout,
+                httpApp,
+                errorHandler,
+                onWriteFailure))
       }
 
     StreamForking.forking(streams, maxConcurrency)
   }
+
+  // private[internal] def reachedEndError[F[_]: Sync](
+  //     socket: Socket[F],
+  //     idleTimeout: Duration,
+  //     receiveBufferSize: Int): Stream[F, Byte] =
+  //   Stream.repeatEval(socket.read(receiveBufferSize, durationToFinite(idleTimeout))).flatMap {
+  //     case None =>
+  //       Stream.raiseError(new EOFException("Unexpected EOF - socket.read returned None") with NoStackTrace)
+  //     case Some(value) => Stream.chunk(value)
+  //   }
+
+  private[internal] def upgradeSocket[F[_]: Concurrent: ContextShift](
+      socketInit: Socket[F],
+      tlsInfoOpt: Option[(TLSContext, TLSParameters)],
+      logger: Logger[F]
+  ): Resource[F, Socket[F]] =
+    tlsInfoOpt.fold(socketInit.pure[Resource[F, *]]) { case (context, params) =>
+      context
+        .server(socketInit, params, { (s: String) => logger.trace(s) }.some)
+        .widen[Socket[F]]
+    }
+
+  private[internal] def runApp[F[_]: Concurrent: Timer](
+      incoming: Stream[F, Byte],
+      maxHeaderSize: Int,
+      requestHeaderReceiveTimeout: Duration,
+      httpApp: HttpApp[F],
+      errorHandler: Throwable => F[Response[F]]): F[(Request[F], Response[F], Stream[F, Byte])] =
+    for {
+      tup <- Parser.Request.parser(maxHeaderSize, durationToFinite(requestHeaderReceiveTimeout))(
+        incoming)
+      (req, rest) = tup
+      resp <- httpApp
+        .run(req)
+        .handleErrorWith(errorHandler)
+        .handleError(_ => serverFailure.covary[F])
+    } yield (req, resp, rest)
+
+  private[internal] def send[F[_]: Sync](socket: Socket[F])(
+      request: Option[Request[F]],
+      resp: Response[F],
+      idleTimeout: Duration,
+      onWriteFailure: (Option[Request[F]], Response[F], Throwable) => F[Unit]): F[Unit] =
+    Encoder
+      .respToBytes[F](resp)
+      .through(socket.writes(durationToFinite(idleTimeout)))
+      .compile
+      .drain
+      .attempt
+      .flatMap {
+        case Left(err) => onWriteFailure(request, resp, err)
+        case Right(()) => Sync[F].pure(())
+      }
+
+  private[internal] def postProcessResponse[F[_]: Timer: Monad](
+      req: Request[F],
+      resp: Response[F]): F[Response[F]] = {
+    val reqHasClose = req.headers.exists {
+      // We know this is raw because we have not parsed any headers in the underlying alg.
+      // If Headers are being parsed into processed for in ParseHeaders this is incorrect.
+      case Header.Raw(name, values) => name == connectionCi && values.contains(closeCi.value)
+      case _ => false
+    }
+    val connection: Connection =
+      if (reqHasClose) close
+      else keepAlive
+    for {
+      date <- HttpDate.current[F].map(Date(_))
+    } yield resp.withHeaders(Headers.of(date, connection) ++ resp.headers)
+  }
+
+  private[internal] def runConnection[F[_]: Concurrent: Timer](
+      socket: Socket[F],
+      logger: Logger[F],
+      idleTimeout: Duration,
+      receiveBufferSize: Int,
+      maxHeaderSize: Int,
+      requestHeaderReceiveTimeout: Duration,
+      httpApp: HttpApp[F],
+      errorHandler: Throwable => F[org.http4s.Response[F]],
+      onWriteFailure: (Option[Request[F]], Response[F], Throwable) => F[Unit]
+  ): Stream[F, Nothing] = {
+    val _ = logger
+    Stream
+      .unfoldLoopEval(socket.reads(receiveBufferSize, durationToFinite(idleTimeout)))(s =>
+        runApp(s, maxHeaderSize, requestHeaderReceiveTimeout, httpApp, errorHandler).attempt.map {
+          case Right((req, resp, rest)) => (Right((req, resp)), Some(rest))
+          case Left(e) => (Left(e), None)
+        })
+      .evalMap {
+        case Right((req, resp)) =>
+          postProcessResponse(req, resp).map(resp => (req, resp).asRight[Throwable])
+        case other => other.pure[F]
+      }
+      .evalTap {
+        case Right((request, response)) =>
+          send(socket)(Some(request), response, idleTimeout, onWriteFailure)
+        case Left(err) =>
+          err match {
+            case req: Parser.Request.ReqPrelude.ParsePreludeError
+                if req == Parser.Request.ReqPrelude.emptyStreamError =>
+              Applicative[F].unit
+            case err =>
+              errorHandler(err)
+                .handleError(_ => serverFailure.covary[F])
+                .flatMap(send(socket)(None, _, idleTimeout, onWriteFailure))
+          }
+      }
+      .takeWhile {
+        case Left(_) => false
+        case Right((req, resp)) =>
+          !(
+            req.headers.get(Connection).exists(_.hasClose) ||
+              resp.headers.get(Connection).exists(_.hasClose)
+          )
+      }
+      .drain ++ Stream.eval_(socket.close)
+  }
+
 }