+++ server/src/test/scala/org/http4s/server/RouterSuite.scala
@@ -80,7 +80,7 @@ class RouterSuite extends Http4sSuite {
           b =!= "bee" && b =!= "one" && res.status === NotFound
         }
       }
-      .assertEquals(true)
+      .assert
   }
 
   test("support root mappings") {
@@ -118,6 +118,6 @@ class RouterSuite extends Http4sSuite {
     val router = Router[IO]("/foo" -> notFound)
     router(Request[IO](uri = uri"/bar")).value
       .map(_ == Option.empty[Response[IO]])
-      .assertEquals(true)
+      .assert
   }
 }