+++ tests/src/test/scala/org/http4s/headers/AllowSuite.scala
@@ -14,19 +14,10 @@
  * limitations under the License.
  */
 
-package org.http4s
-package headers
+package org.http4s.headers
 
-import org.http4s.ServerSentEvent._
-import org.scalacheck._
-import org.scalacheck.Arbitrary._
+import org.http4s.laws.discipline.ArbitraryInstances._
 
-class LastEventIdSpec extends HeaderLaws {
-  implicit val arbLastEventId: Arbitrary[`Last-Event-Id`] =
-    Arbitrary(for {
-      id <- arbitrary[String]
-      if !id.contains("\uffff")
-    } yield `Last-Event-Id`(EventId(id)))
-
-  checkAll("Last-Event-Id", headerLaws(`Last-Event-Id`))
+class AllowSuite extends MHeaderLaws {
+  checkAll("Allow", headerLaws(Allow))
 }